requirements.txt
amqp==5.3.1
asgiref==3.8.1
astroid==2.15.8
async-timeout==5.0.1
autoflake==2.2.0
bandit==1.7.5
billiard==4.2.1
black==22.8.0
celery==5.5.1
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dill==0.4.0
distlib==0.3.9
dj-database-url==2.0.0
Django==4.2
django-redis==5.4.0
djangorestframework==3.14.0
dparse==0.6.4
drf-yasg==1.21.10
filelock==3.18.0
flake8==6.0.0
gitdb==4.0.12
GitPython==3.1.44
gunicorn==20.1.0
identify==2.6.10
idna==3.10
inflection==0.5.1
isort==5.12.0
kombu==5.5.3
lazy-object-proxy==1.11.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==21.3
pathspec==0.12.1
pbr==6.1.1
platformdirs==4.3.7
pre-commit==3.3.3
prompt_toolkit==3.0.51
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.19.1
pylint==2.17.7
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safety==2.3.5
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
stevedore==5.4.1
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
virtualenv==20.30.0
wcwidth==0.2.13
whitenoise==6.4.0
wrapt==1.17.2
